Monad comprehensions without RebindableSyntax.
This version of the embedding relies on a GHC patch that modifies the
desugaring of monad comprehensions. Comprehension guards are not
desugared into Control.Monad.guard but into an overloaded version in a
new class Guardable. Guardable also abstracts over the boolean and unit
type used in guard to allow a deep embedding to use other types.